Begin Metric Consolidation
